moved tests to /tests directory, and pulled interfaces into their own files
