+++ src/eosjs-api.ts
@@ -6,8 +6,10 @@
 
 "use strict";
 
-import { Abi, GetInfoResult, JsonRpc, PushTransactionArgs } from "./eosjs-jsonrpc";
+import { JsonRpc } from "./eosjs-jsonrpc";
 import * as ser from "./eosjs-serialize";
+import { AuthorityProvider, AbiProvider, BinaryAbi, CachedAbi, SignatureProvider } from "./api-interfaces";
+import { Abi, GetInfoResult, PushTransactionArgs } from "./jsonrpc-interfaces";
 
 // tslint:disable-next-line
 const abiAbi = require('../src/abi.abi.json');
@@ -16,65 +18,6 @@ const transactionAbi = require('../src/transaction.abi.json');
 
 /** Reexport `eosjs-serialize` */
 export const serialize = ser;
-
-/** Arguments to `getRequiredKeys` */
-export interface AuthorityProviderArgs {
-    /** Transaction that needs to be signed */
-    transaction: any;
-
-    /** Public keys associated with the private keys that the `SignatureProvider` holds */
-    availableKeys: string[];
-}
-
-/** Get subset of `availableKeys` needed to meet authorities in `transaction` */
-export interface AuthorityProvider {
-    /** Get subset of `availableKeys` needed to meet authorities in `transaction` */
-    getRequiredKeys: (args: AuthorityProviderArgs) => Promise<string[]>;
-}
-
-/** Retrieves raw ABIs for a specified accountName */
-export interface AbiProvider {
-    getRawAbi: (accountName: string) => Promise<BinaryAbi>;
-}
-
-/** Structure for the raw form of ABIs */
-export interface BinaryAbi {
-    account_name: string;
-    abi: Uint8Array;
-}
-
-/** Arguments to `sign` */
-export interface SignatureProviderArgs {
-    /** Chain transaction is for */
-    chainId: string;
-
-    /** Public keys associated with the private keys needed to sign the transaction */
-    requiredKeys: string[];
-
-    /** Transaction to sign */
-    serializedTransaction: Uint8Array;
-
-    /** ABIs for all contracts with actions included in `serializedTransaction` */
-    abis: BinaryAbi[];
-}
-
-/** Signs transactions */
-export interface SignatureProvider {
-    /** Public keys associated with the private keys that the `SignatureProvider` holds */
-    getAvailableKeys: () => Promise<string[]>;
-
-    /** Sign a transaction */
-    sign: (args: SignatureProviderArgs) => Promise<string[]>;
-}
-
-/** Holds a fetched abi */
-export interface CachedAbi {
-    /** abi in binary form */
-    rawAbi: Uint8Array;
-
-    /** abi in structured form */
-    abi: Abi;
-}
 
 export class Api {
     /** Issues RPC calls */
