+++ src/index.ts
@@ -5,4 +5,3 @@ import SignatureProvider from "./eosjs-jssig";
 import * as Serialize from "./eosjs-serialize";
 
 export { Api, SignatureProvider, Rpc, Serialize };
-export default { Api, SignatureProvider, Rpc, Serialize };
