+++ src/tests/eosjs-jssig.test.ts
@@ -1,6 +1,6 @@
 import * as ecc from "eosjs-ecc";
 
-import JsSignatureProvider from "./eosjs-jssig";
+import JsSignatureProvider from "../eosjs-jssig";
 
 describe("JsSignatureProvider", () => {
     const privateKeys = ["key1", "key2", "key3"];
