+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "target": "es5",
-        "module": "commonjs",
+        "module": "CommonJS",
         "outDir": "dist",
         "sourceMap": true,
         "noImplicitAny": true,
